feat(market_watch)!: use string type for ins_code index
tests: fix assert_dict_type to not use pydantic
pydantic v2.0 has removed create_model_from_typeddict

+++ tsetmc/instruments.py
@@ -136,8 +136,8 @@ class _ETF(_TypedDict):
     insCode: str
     deven: int
     hEven: int
-    pRedTran: int
-    pSubTran: int
+    pRedTran: float
+    pSubTran: float
     iClose: int
 
 
@@ -1023,7 +1023,7 @@ class _Instrument(_TypedDict, total=False):
 
 class _ShareHolderCompany(_TypedDict):
     instrument: _Instrument
-    numberOfShares: int
+    numberOfShares: float
     perOfShares: float
